# [Core] Enable multiple profiler consumers and add a timeline/tracing profiler