--- a/sources/core/Stride.Core.Tests/TestProfiler.cs
+++ b/sources/core/Stride.Core.Tests/TestProfiler.cs
@@ -1,12 +1,16 @@
 // Copyright (c) .NET Foundation and Contributors (https://dotnetfoundation.org/ & https://stride3d.net) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
-using Xunit;
-using Stride.Core.Diagnostics;
 using System.Threading;
+using System.Threading.Channels;
+using System.Threading.Tasks;
+using Stride.Core.Diagnostics;
+using Xunit;
 
 namespace Stride.Core.Tests
 {
@@ -107,7 +111,7 @@ namespace Stride.Core.Tests
 
 
         [Fact]
-        public void TestWitAttributes()
+        public void TestWithAttributes()
         {
             Profiler.Reset();
             const int timeToWait = 10;
@@ -131,6 +135,78 @@ namespace Stride.Core.Tests
             watcher.Finish();
         }
 
+        [Fact]
+        public async void TestSubscribersReceiveEvents()
+        {
+            const int subscriberCount = 5;
+            const int eventCount = 5;
+
+            var subscribers = new TestEventReader[subscriberCount];
+
+            Profiler.DisableAll();
+
+            for(int i = 0; i < subscriberCount; i++)
+            {
+                // EventReaders will Unsubscribe themselves after reading 'eventCount' events.
+                subscribers[i] = new TestEventReader(eventsToRead: eventCount);
+                subscribers[i].Subscribe();
+            }
+
+            Profiler.MinimumProfileDuration = TimeSpan.Zero;
+            Profiler.EnableAll();
+
+            for(int i = 0; i < eventCount; i++)
+            {
+                using var _ = Profiler.Begin(TestKey);
+            }
+
+            var results = await Task.WhenAll(subscribers.Select(async x => await x.ReadAll()));            
+
+            Assert.All(results, x => Assert.Equal(eventCount, x.Count));
+        }
+
+        [Fact]
+        public async void TestConcurrentUse()
+        {
+            const int subscriberCount = 100;
+            const int eventCount = 5;
+
+            using CancellationTokenSource cts = new CancellationTokenSource();
+
+            Task[] eventGenerators = Enumerable.Range(0, 4).Select(x => Task.Run(() =>
+            {
+                while (true)
+                {
+                    if (cts.IsCancellationRequested)
+                        return;
+
+                    using var _ = Profiler.Begin(TestKey);
+                    Thread.Sleep(1);
+                }
+            })).ToArray();
+
+            Profiler.MinimumProfileDuration = TimeSpan.Zero;
+            Profiler.EnableAll();
+
+            var subscribers = new ConcurrentBag<TestEventReader>();
+
+            Parallel.For(0, subscriberCount, (i) =>
+            {
+                // EventReaders will Unsubscribe themselves after reading 'eventCount' events.
+                var reader = new TestEventReader(eventsToRead: eventCount);
+                reader.Subscribe();
+                subscribers.Add(reader);
+                Thread.Sleep(1);
+            });
+            
+            var results = await Task.WhenAll(subscribers.Select(async x => await x.ReadAll()));            
+            
+            cts.Cancel();
+
+            Task.WaitAll(eventGenerators);
+
+            Assert.All(results, x => Assert.Equal(eventCount, x.Count));
+        }
 
         private static Regex matchElapsed = new Regex(@"Elapsed = ([\d.]+)", RegexOptions.CultureInvariant);
 
@@ -221,6 +297,47 @@ namespace Stride.Core.Tests
             };
             GlobalLogger.GlobalMessageLogged += watcher.LogAction;
             return watcher;
+        }
+
+        private class TestEventReader
+        { 
+
+            ChannelReader<ProfilingEvent> reader;
+
+            public List<ProfilingEvent> Events;
+            private int eventsToRead;
+
+            public TestEventReader(int eventsToRead)
+            {
+                this.eventsToRead = eventsToRead;
+            }
+
+            public void Subscribe()
+            {
+                Events = new List<ProfilingEvent>();
+                reader = Profiler.Subscribe();
+            }
+
+            public void Unsubscribe()
+            {
+                Profiler.Unsubscribe(reader);
+            }
+
+            public async Task<List<ProfilingEvent>> ReadAll()
+            {
+                await foreach (var item in reader.ReadAllAsync())
+                {
+                    Events.Add(item);
+                    if (Events.Count == eventsToRead)
+                    {
+                        Unsubscribe();
+                        break;
+                    }
+                }
+
+                return Events;
+            }
+
         }
     }
 }