… sources/core/Stride.Core/Extensions/TimeSpanExtensions.cs
… sources/core/Stride.Core/Threading/ConcurrentCollector.cs
… sources/core/Stride.Core/Threading/Dispatcher.cs
… sources/editor/Stride.Assets.Presentation/SceneEditor/MaterialFilterRenderFeature.cs
… sources/engine/Stride.Audio/StreamedBufferSoundSource.cs
… sources/engine/Stride.Engine/Engine/EntityManager.cs
… sources/engine/Stride.Engine/Profiling/GameProfilingSorting.cs
… sources/engine/Stride.Engine/Profiling/GameProfilingSystem.cs
… sources/engine/Stride.Engine/Rendering/Compositing/ForwardRenderer.cs
… sources/engine/Stride.Engine/Rendering/Compositing/GraphicsCompositor.cs
… sources/engine/Stride.Engine/Rendering/Compositing/SceneCameraRenderer.cs
… sources/engine/Stride.Rendering/Rendering/Compositing/SceneRendererCollection.cs
… sources/engine/Stride.Rendering/Rendering/InstancingRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/Lights/ForwardLightingRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/Materials/MaterialRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/MeshRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/MeshVelocityRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/RenderSystem.cs
… sources/engine/Stride.Rendering/Rendering/RendererBase.cs
… sources/engine/Stride.Rendering/Rendering/RendererCoreBase.cs
… sources/engine/Stride.Rendering/Rendering/RootEffectRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/RootRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/Shadows/ShadowMapRenderer.cs
… sources/engine/Stride.Rendering/Rendering/SkinningRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/TransformRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/VisibilityGroup.cs
… sources/engine/Stride.Video/VideoInstance.cs