# PR Details

## Description

This PR adds a profiler outputting data in chrome://tracing format (for now) and changes `Profiler` to make that possible.

`ChromeTracingProfileWriter` consumes events and dumps them to a file which can then be analyzed with tools like chrome://tracing, [perfetto](https://ui.perfetto.dev) or [speedscope](https://speedscope.app).

### Usage
```cs
//Start
ChromeTracingProfileWriter.Start(string outputPath, bool indentOutput = false)
//Stop
ChromeTracingProfiler.Stop()

//To add another output format/client and receive events use
Profiler.Subscribe() and Profiler.Unsubscribe()
```

## Related Issue

N/A

## Motivation and Context

I had trouble pinpointing performance issues with both the existing Profiler and external tools like dotTrace, so I started experimenting with adding something in the vein of chrome://tracing or [tracy](https://github.com/wolfpld/tracy). In particular the exisiting `Profiler` is quite heavy-weight and doesn't play well with multithreading, while everything happening inside the `ThreadPool` / `Dispatcher` is difficult to understand with sampling profilers as a lot of work happens in anonymous functions.

The `TraceProfiler` lets you see work on all threads and lets you look at individual frames to analyze spikes instead of just averages. For me this was also very helpful in understanding the structure of the engine.

An example trace from a rendering stresstest opened in [perfetto](https://ui.perfetto.dev):
![grafik](https://github.com/stride3d/stride/assets/8515865/f1953a3a-ec19-4cf0-9866-497e2a04a0cb)

The example trace:
[tracing_example.json.gz](https://github.com/stride3d/stride/files/12643239/tracing_example.json.gz)

## Open Questions / Todo

- [x] Decide what to do with this. I've already found this very useful compared with the existing Profiler, so I'd like to see it added, but of course having multiple internal profilers is not ideal. Merging it with `Profiler` seemed quite challenging, but maybe it's a bit easier after the refactoring in #1774 and I should take another look.
- [x] Add tests.
- [x] Update outdated comments and add new ones where needed.
- [x] Verify performance.
- [ ] Update docs.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [x] My change requires a change to the documentation.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.