@manio143 2023-09-18
That is very interesting. I want to see this as an opt-in. And given the implementation is not large I think we could merge it into the source code.
What do you think about implementing a hook into the existing Profiler?
Something like
```csharp
public delegate void EventProcessedHandler(ref ProfilingEvent profilingEvent, ProfilingEventType eventType);
public static event EventProcessedHandler EventProcessed;

// at the end of Profiler.ProcessEvent
EventProcessed?.Invoke(ref profilingEvent, eventType);

// In your implementation a new static method
TraceProfiler.StartCollectingEvents() => Profiler.EventProcessed += OnProfilingEventProcessed;
TraceProfiler.StopCollectingEvents() => Profiler.EventProcessed -= OnProfilingEventProcessed;
```
And then just add more `ProfilingKey`s to places where they were missing - like you did with your implementation specifically?

And if there's something in the current Profiler implementation that prevents it from being that simple - maybe we can look at why is that?