@manio143 2023-09-18
We can easily add caller info into the Profiler. Other than that I see mainly that you're adding thread info and maintaining a per thread stack in order to stop the Stopwatch in order for Begin/End pairs (already done by ProfilingState).
The last difference - instead of using a per frame ring buffer to store events you use async channels to stream them to a file.