@jazzay 2023-09-18
I'm curious how we ensure profiling call sites get stripped for release/shipping builds?