@froce 2023-09-18
> And then just add more `ProfilingKey`s to places where they were missing - like you did with your implementation specifically?
> 
> And if there's something in the current Profiler implementation that prevents it from being that simple - maybe we can look at why is that?

My first concern here would be performance. How big of a deal that is really depends on what you're profiling: A few hundred scopes spread around should be fine, thousands of scopes in a small function that gets called by worker threads is perf death by synchronization overhead.

As a quick best-cast vs worst-case example this is just collecting 100k events, without any output, either in a regular `for` loop or in a `Parallel.For`:
``` ini

BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.19045.3324)
AMD Ryzen 5 3600, 1 CPU, 12 logical and 6 physical cores
.NET SDK=8.0.100-preview.7.23376.3
  [Host]     : .NET 6.0.14 (6.0.1423.7309), X64 RyuJIT AVX2
  Job-JPLRAA : .NET 6.0.14 (6.0.1423.7309), X64 RyuJIT AVX2

InvocationCount=1  UnrollFactor=1  

```
|                 Method |      Mean |     Error |    StdDev | Ratio | RatioSD |
|----------------------- |----------:|----------:|----------:|------:|--------:|
|         StrideProfiler | 42.977 ms | 0.7726 ms | 0.6032 ms |  3.85 |    0.35 |
|                  Trace | 11.706 ms | 0.6019 ms | 1.7556 ms |  1.00 |    0.00 |
| StrideProfilerParallel | 77.534 ms | 1.5432 ms | 2.8985 ms |  6.86 |    0.75 |
|          TraceParallel |  2.490 ms | 0.1216 ms | 0.3566 ms |  0.22 |    0.05 |

I also thought that adding lower level profiling keys would just clutter the existing profiling UI.
Maybe adding a `ProfilingKeyFlags.TracingOnly` would be a good option.

I think my other issues were really about `Profiler` itself and shouldn't affect this PR, your PR has already improved it and the rest can wait for a follow-up.
I'm also not a fan of how adding/removing `ProfilingKeys` even for temporary profiling is needed, but it's a small price