@manio143 2023-09-18
> I'm curious how we ensure profiling call sites get stripped for release/shipping builds?

They don't need to. When the profiler is off the JIT can inline the check and skip the calls to it quite well.

If the Profiler perf is not great it would be good to determine why and see if we can improve it. Maybe it would make sense to gut it even more and use something closer to the runtime? What we need is basically a simple solution to start/stop scopes and gather perf info and output it to some stream once per frame - either consumed by in game UI or by external tools or written to a file, etc.