@jazzay 2023-09-18
> They don't need to. When the profiler is off the JIT can inline the check and skip the calls to it quite well.
In the current implementation of TracingProfiler/Scope I don't see any is active checks, so how would it avoid that work?

Additionally. what about AOT platforms like iOS, etc?