@froce 2023-10-04
I think I've finally figured out what I want to do with this. In a first step, replace the Profiler internal buffers with Channels to get rid of the locking and to get support for multiple consumers. Then I can convert the TraceProfiler callsites to the existing functions/datatypes and add the JSON writer as a second consumer. The only functional change I would want to make here is to stop sending `Begin` events, they just get filtered out on the receiving side (`GameProfilingSystem`) anyway, while basically doubling the event count for no benefit.

Then, as a second PR, we could take another look at the existing API and datatypes:
- Add caller info
- I haven't measured the impact, but `ProfilingState` and `ProfilingEvent` are currently both 296 bytes, that's bound to have a perf impact. `Attributes` make up 144 bytes in each and they are **never** used, Custom messages are 112 bytes, but at least sometimes used. So we can probably slim that down, or split things up, so only what's used is actually recorded.
- `Mark` events are very rarely used and could basically be replaced with 0 duration timespans. Maybe we should add an API for counters instead
- Per frame grouping of events (okay, maybe that's a good use for markers)

And then, as a last step, take another look at GPU profiling. Usually I would expect GPU information per queue (on modern APIs), but I don't think that's currently exposed. But I might very well be wrong about that, didn't want to dig to deeply into it for now.

As an alternative to rolling our own we could also look at integrating [https://github.com/clibequilibrium/Tracy-CSharp](https://github.com/clibequilibrium/Tracy-CSharp), it's definitely well optimized and feature rich, but ofc it's another (non-C#) dependency (+ imgui) and I don't know what limitations the .NET integration might have.