@froce 2023-10-04
> Other than that I see mainly that you're adding thread info and maintaining a per thread stack in order to stop the Stopwatch in order for Begin/End pairs (already done by ProfilingState).

The stack is only there to support profiling across function boundaries, with ProfilingState you have to pass it along yourself somehow, if begin and end are not in the same scope. (That also means my attempt at limiting event capturing by stack depth was not working, oops. I will remove it, did not seem useful in practice anyway.)

> > They don't need to. When the profiler is off the JIT can inline the check and skip the calls to it quite well.
> 
> In the current implementation of TracingProfiler/Scope I don't see any is active checks, so how would it avoid that work?
> 
> Additionally. what about AOT platforms like iOS, etc?

Yeah, no checks in the current version. I will add runtime checks / use the existing ones in Profiler/ProfilingKey. We could add a compile time way of disabling profiling entirely, if it's needed, but I'd prefer having the profiler available in all builds by default.