@manio143 2023-10-04
I think we might be able to separate the message logging from the profiler. There isn't really a good reason for it to be tied together (except being enabled when profiling is enabled) and it will simplify things a bit.

For having variants of the same profiling key attributes should be used.
And for things like the GC counts we may need a counter primitive as you mentioned instead.